+++ NEWS.md
@@ -1,7 +1,7 @@
 # odbc (development version)
 
-* Added functions `odbcConfigureDrivers()`, `odbcConfigureSystemDSN()`, and
-  `odbcConfigureUserDSN()` to provide a shorthand for 
+* Added functions `odbcEditDrivers()`, `odbcEditSystemDSN()`, and
+  `odbcEditUserDSN()` to provide a shorthand for 
   `file.edit(odbcListConfig()[[i]])`. The helpers only support macOS and 
   Linux and will raise an error on Windows (@simonpcouch, #827).